+++ secureli/actions/action.py
@@ -11,14 +11,14 @@ from secureli.repositories.secureli_config import (
 from secureli.repositories.settings import SecureliRepository, TelemetrySettings
 from secureli.services.language_analyzer import LanguageAnalyzerService
 from secureli.services.language_support import (
-    LanguageMetadata,
     LanguageSupportService,
 )
-from secureli.services.scanner import ScannerService, ScanMode
+from secureli.services.scanner import ScannerService
 from secureli.services.updater import UpdaterService
 
 from secureli.shared.models.actions import VerifyOutcome, VerifyResult
-from secureli.shared.models.language import LanguageNotSupportedError
+from secureli.shared.models.language import LanguageMetadata, LanguageNotSupportedError
+from secureli.shared.models.scan import ScanMode
 from secureli.utilities.formatter import format_sentence_list
