+++ secureli/actions/setup.py
@@ -1,6 +1,5 @@
 import jinja2
-
-from secureli.services.language_support import supported_languages
+from secureli.shared.consts.language import supported_languages
 
 
 class SetupAction:
